Fix CSS on published pages and previews
CSS was not being included for published pages and previews. This change
fixes it by including CSS in components' SSR bundles, which are used for
static page rendering. This shouldn't change the fact that component JS
bundles does not contain CSS because the CSS will be pre-rendered and
embedded in the page HTML.